refactor(discord): bootstrap の agent 生成ロジックを agents.ts へ分離する
