+++ apps/discord/src/bootstrap.test.ts
@@ -1,20 +1,10 @@
 /* oxlint-disable max-lines, max-lines-per-function -- bootstrap の DI 結合テストはケース数に応じて長くなるため許容 */
 import { afterEach, describe, expect, mock, test } from "bun:test";
-import { mkdtempSync, readFileSync } from "fs";
-import os from "os";
-import { join } from "path";
 
 import { createMockLogger } from "@vicissitude/shared/test-helpers";
 import type { DueReminder } from "@vicissitude/shared/types";
 
-import {
-	buildEmailCheckPreFilter,
-	createDiscordAgents,
-	createWebConversationAgent,
-	createMetrics,
-	resolveBootstrapRoot,
-} from "./bootstrap.ts";
-import { createStoreLayer } from "./bootstrap/layers.ts";
+import { buildEmailCheckPreFilter, createMetrics, resolveBootstrapRoot } from "./bootstrap.ts";
 import type { AppConfig } from "./config.ts";
 
 function createTestConfig(overrides?: Partial<AppConfig>): AppConfig {
@@ -79,433 +69,6 @@ describe("resolveBootstrapRoot", () => {
 		);
 
 		expect(root).toBe("/tmp/vicissitude-root");
-	});
-});
-
-describe("createDiscordAgents", () => {
-	test("Discord guild agent に core と discord MCP を渡す", () => {
-		const config = createTestConfig();
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: { DATA_DIR: "/data/core" },
-				discordEnvironment: { DISCORD_TOKEN: "token", DATA_DIR: "/data/discord" },
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			profile: {
-				mcpServers: Record<string, { type: string; environment?: Record<string, string> }>;
-				skillPermission: Record<string, string>;
-			};
-			sessionPort: { config: { skillPaths?: string[] } };
-		};
-
-		expect(Object.keys(agent.profile.mcpServers).toSorted()).toEqual(["core", "discord"]);
-		expect(agent.profile.mcpServers.core?.environment?.AGENT_ID).toBe("discord:123456789");
-		expect(agent.profile.mcpServers.discord?.environment?.AGENT_ID).toBe("discord:123456789");
-		expect(agent.profile.mcpServers.discord?.environment?.DISCORD_TOKEN).toBe("token");
-		expect(agent.profile.skillPermission).toEqual({ "*": "deny" });
-		expect(agent.sessionPort.config.skillPaths).toEqual(["/app/context/skills/discord"]);
-	});
-
-	test("Minecraft 有効時は Discord agent に minecraft skill を許可する", () => {
-		const config = createTestConfig({
-			minecraft: {
-				host: "localhost",
-				port: 25565,
-				username: "hua",
-				authMode: "offline",
-				mcpPort: 3001,
-				viewerPort: 3007,
-			},
-		});
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: {},
-				discordEnvironment: { DISCORD_TOKEN: "token", DATA_DIR: "/data/discord" },
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			profile: {
-				builtinTools: Record<string, boolean>;
-				skillPermission: Record<string, string>;
-			};
-			sessionPort: { config: { skillPaths?: string[] } };
-		};
-
-		expect(agent.profile.builtinTools.skill).toBe(true);
-		expect(agent.profile.skillPermission).toEqual({
-			"*": "deny",
-			minecraft: "allow",
-		});
-		expect(agent.sessionPort.config.skillPaths).toEqual(["/app/context/skills/discord"]);
-	});
-
-	test("Discord DM agent は DM scopeId と agentId で作成される", () => {
-		const config = createTestConfig();
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:dm:999888777", scopeId: "discord:dm:999888777" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: { DATA_DIR: "/data/core" },
-				discordEnvironment: { DISCORD_TOKEN: "token", DATA_DIR: "/data/discord" },
-			},
-		);
-		const agent = agents.get("discord:dm:999888777") as unknown as {
-			profile: {
-				mcpServers: Record<string, { environment?: Record<string, string> }>;
-			};
-		};
-
-		expect(agent.profile.mcpServers.core?.environment?.AGENT_ID).toBe("discord:dm:999888777");
-		expect(agent.profile.mcpServers.discord?.environment?.AGENT_ID).toBe("discord:dm:999888777");
-	});
-
-	test("heartbeat agent は shellWorkspace 有効時でも deps の OpenCode 設定だけを使う", () => {
-		const config = createTestConfig({
-			shellAgent: {
-				enabled: true,
-				agent: {
-					providerId: "shell-provider",
-					modelId: "shell-model",
-				},
-				backgroundSubagents: true,
-				dataDir: "/tmp/shell-workspaces",
-			},
-		});
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:heartbeat:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: {},
-				discordEnvironment: {},
-				opencode: {
-					providerId: "heartbeat-provider",
-					modelId: "heartbeat-model",
-					temperature: 0.2,
-				},
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			profile: {
-				model: { providerId: string; modelId: string };
-				builtinTools: Record<string, boolean>;
-				skillPermission: Record<string, string>;
-				opencodeAgents?: unknown;
-			};
-			sessionPort: {
-				config: {
-					temperature?: number;
-					skillPaths?: string[];
-					directory?: string;
-					environment?: Record<string, string>;
-				};
-			};
-		};
-
-		expect(agent.profile.model).toEqual({
-			providerId: "heartbeat-provider",
-			modelId: "heartbeat-model",
-		});
-		expect(agent.sessionPort.config.temperature).toBe(0.2);
-		expect(agent.profile.builtinTools.skill).toBe(false);
-		expect(agent.profile.builtinTools.bash).toBe(false);
-		expect(agent.profile.builtinTools.task).toBe(false);
-		expect(agent.profile.skillPermission).toEqual({ "*": "deny" });
-		expect(agent.profile.opencodeAgents).toBeUndefined();
-		expect(agent.sessionPort.config.skillPaths).toEqual(["/app/context/skills/discord"]);
-		expect(agent.sessionPort.config.directory).toBeUndefined();
-		expect(agent.sessionPort.config.environment).toBeUndefined();
-	});
-
-	test("Minecraft 有効時は heartbeat agent でも minecraft skill を許可する", () => {
-		const config = createTestConfig({
-			minecraft: {
-				host: "localhost",
-				port: 25565,
-				username: "hua",
-				authMode: "offline",
-				mcpPort: 3001,
-				viewerPort: 3007,
-			},
-		});
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:heartbeat:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: {},
-				discordEnvironment: {},
-				opencode: {
-					providerId: "heartbeat-provider",
-					modelId: "heartbeat-model",
-					temperature: 0.2,
-				},
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			profile: {
-				builtinTools: Record<string, boolean>;
-				skillPermission: Record<string, string>;
-				opencodeAgents?: unknown;
-			};
-		};
-
-		expect(agent.profile.builtinTools.skill).toBe(true);
-		expect(agent.profile.skillPermission).toEqual({
-			"*": "deny",
-			minecraft: "allow",
-		});
-		expect(agent.profile.opencodeAgents).toBeUndefined();
-	});
-
-	test("shellWorkspace の GitHub token は Git credential helper 付きで OpenCode に渡す", () => {
-		const config = createTestConfig({
-			shellAgent: {
-				enabled: true,
-				agent: {
-					providerId: "shell-provider",
-					modelId: "shell-model",
-				},
-				environment: {
-					GH_TOKEN: "github-token",
-					GITHUB_TOKEN: "github-token",
-				},
-				dataDir: "/tmp/shell-workspaces",
-			},
-		});
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: {},
-				discordEnvironment: {},
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			sessionPort: { config: { environment?: Record<string, string> } };
-		};
-
-		expect(agent.sessionPort.config.environment?.GH_TOKEN).toBe("github-token");
-		expect(agent.sessionPort.config.environment?.GIT_CONFIG_COUNT).toBe("1");
-		expect(agent.sessionPort.config.environment?.GIT_CONFIG_KEY_0).toBe(
-			"credential.https://github.com.helper",
-		);
-		expect(agent.sessionPort.config.environment?.GIT_CONFIG_VALUE_0).toContain("GH_TOKEN");
-		expect(agent.sessionPort.config.environment?.GIT_CONFIG_VALUE_0).not.toContain("github-token");
-	});
-
-	test("shellWorkspace の Git identity は workspace 内の gitconfig として渡す", () => {
-		const dataDir = mkdtempSync(join(os.tmpdir(), "vicissitude-shell-workspace-"));
-		const config = createTestConfig({
-			shellAgent: {
-				enabled: true,
-				agent: {
-					providerId: "shell-provider",
-					modelId: "shell-model",
-				},
-				git: {
-					userName: "ふあ",
-					userEmail: "282728168+agenthua@users.noreply.github.com",
-				},
-				dataDir,
-			},
-		});
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: {},
-				discordEnvironment: {},
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			sessionPort: { config: { environment?: Record<string, string> } };
-		};
-		const gitConfigPath = join(
-			dataDir,
-			"opencode",
-			"discord_123456789",
-			".config",
-			"git",
-			"config",
-		);
-		const gitConfig = readFileSync(gitConfigPath, "utf8");
-
-		expect(agent.sessionPort.config.environment?.GIT_CONFIG_GLOBAL).toBe(gitConfigPath);
-		expect(gitConfig).toContain('name = "ふあ"');
-		expect(gitConfig).toContain('email = "282728168+agenthua@users.noreply.github.com"');
-		expect(gitConfig).toContain('[credential "https://github.com"]');
-	});
-
-	test("backgroundSubagents 有効時は OpenCode 実験フラグを渡す", () => {
-		const config = createTestConfig({
-			shellAgent: {
-				enabled: true,
-				agent: {
-					providerId: "shell-provider",
-					modelId: "shell-model",
-				},
-				backgroundSubagents: true,
-				dataDir: "/tmp/shell-workspaces",
-			},
-		});
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: {},
-				discordEnvironment: {},
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			profile: {
-				primaryTools?: string[];
-				builtinTools: Record<string, boolean>;
-				opencodeAgents?: Record<string, { tools?: Record<string, boolean>; permission?: unknown }>;
-			};
-			sessionPort: { config: { environment?: Record<string, string>; skillPaths?: string[] } };
-		};
-
-		expect(agent.profile.builtinTools.skill).toBe(true);
-		expect(agent.profile.builtinTools.task_status).toBe(true);
-		expect(agent.profile.primaryTools).toEqual(["task", "task_status", "skill"]);
-		expect(agent.profile.opencodeAgents?.build?.tools?.skill).toBe(true);
-		expect(agent.profile.opencodeAgents?.build?.permission).toMatchObject({
-			skill: { "*": "deny", "delegate-to-shell-worker": "allow", "self-update": "allow" },
-		});
-		expect(agent.profile.opencodeAgents?.["shell-worker"]?.tools?.skill).toBe(true);
-		expect(agent.profile.opencodeAgents?.["shell-worker"]?.permission).toMatchObject({
-			"*_*": "deny",
-			"core_*": "deny",
-			"discord_*": "deny",
-			"mc-bridge_*": "deny",
-			"minecraft_*": "deny",
-		});
-		expect(agent.sessionPort.config.skillPaths).toEqual([
-			"/app/context/skills/discord",
-			"/app/context/skills/shell-worker",
-		]);
-		expect(agent.sessionPort.config.environment?.OPENCODE_EXPERIMENTAL_BACKGROUND_SUBAGENTS).toBe(
-			"true",
-		);
-	});
-
-	test("shellWorkspace と Minecraft 併用時は primary agent に delegate-to-shell-worker と minecraft skill を許可する", () => {
-		const config = createTestConfig({
-			minecraft: {
-				host: "localhost",
-				port: 25565,
-				username: "hua",
-				authMode: "offline",
-				mcpPort: 3001,
-				viewerPort: 3007,
-			},
-			shellAgent: {
-				enabled: true,
-				agent: {
-					providerId: "shell-provider",
-					modelId: "shell-model",
-				},
-				dataDir: "/tmp/shell-workspaces",
-			},
-		});
-		const { db, sessionStore } = createStoreLayer(config);
-		const agents = createDiscordAgents(
-			config,
-			[{ agentId: "discord:123456789", scopeId: "discord:guild:123456789" }],
-			{
-				db,
-				sessionStore,
-				contextBuilder: { build: () => Promise.resolve("context") },
-				logger: createMockLogger(),
-				appRoot: "/app",
-				coreEnvironment: {},
-				discordEnvironment: {},
-			},
-		);
-		const agent = agents.get("discord:guild:123456789") as unknown as {
-			profile: {
-				primaryTools?: string[];
-				opencodeAgents?: Record<string, { tools?: Record<string, boolean>; permission?: unknown }>;
-			};
-		};
-
-		const build = agent.profile.opencodeAgents?.build as
-			| { tools?: Record<string, boolean>; permission?: { skill?: Record<string, string> } }
-			| undefined;
-		const worker = agent.profile.opencodeAgents?.["shell-worker"] as
-			| { tools?: Record<string, boolean>; permission?: { skill?: Record<string, string> } }
-			| undefined;
-
-		expect(agent.profile.primaryTools).toEqual(["task", "skill"]);
-		expect(build?.tools?.skill).toBe(true);
-		expect(build?.permission?.skill).toEqual({
-			"*": "deny",
-			"delegate-to-shell-worker": "allow",
-			"self-update": "allow",
-			minecraft: "allow",
-		});
-		expect(worker?.permission?.skill).toEqual({
-			"*": "deny",
-			debug: "allow",
-			"skill-creator": "allow",
-		});
-		expect(worker?.permission).toMatchObject({
-			"*_*": "deny",
-			"discord_*": "deny",
-			"minecraft_*": "deny",
-		});
 	});
 });
 
@@ -653,34 +216,5 @@ describe("buildEmailCheckPreFilter 内部分岐", () => {
 		expect(logger.error).toHaveBeenCalled();
 		expect(result.reminders).toEqual([]);
 		expect(result.markExecutedIds).toEqual(["email-check"]);
-	});
-});
-
-describe("createWebConversationAgent", () => {
-	test("Web agent は core MCP のみを持ち AGENT_ID を web:local にする", () => {
-		const config = createTestConfig();
-		const { sessionStore } = createStoreLayer(config);
-		const agent = createWebConversationAgent(config, {
-			sessionStore,
-			contextBuilder: { build: () => Promise.resolve("context") },
-			logger: createMockLogger(),
-			appRoot: "/app",
-			coreEnvironment: { DATA_DIR: "/data/core" },
-			opencodePort: 4103,
-		}) as unknown as {
-			profile: {
-				mcpServers: Record<string, { type: string; environment?: Record<string, string> }>;
-				builtinTools: Record<string, boolean>;
-				skillPermission: Record<string, string>;
-			};
-			sessionPort: { config: { port: number; skillPaths?: string[] } };
-		};
-
-		expect(Object.keys(agent.profile.mcpServers)).toEqual(["core"]);
-		expect(agent.profile.mcpServers.core?.environment?.AGENT_ID).toBe("web:local");
-		expect(agent.profile.builtinTools.skill).toBe(false);
-		expect(agent.profile.skillPermission).toEqual({ "*": "deny" });
-		expect(agent.sessionPort.config.port).toBe(4103);
-		expect(agent.sessionPort.config.skillPaths).toBeUndefined();
 	});
 });
